+++ features/admin.core.v1/models/config.ts
@@ -467,6 +467,10 @@ export interface UIConfigInterface extends CommonUIConfigInterface<FeatureConfig
      * Config if beta tag should be displayed for sms otp for password recovery feature.
      */
     showSmsOtpPwdRecoveryFeatureStatusChip?: boolean;
+    /**
+     * Config to check whether consent is required for trusted apps.
+     */
+    isTrustedAppConsentRequired?: boolean;
 }
 
 /**