+++ modules/i18n/src/models/namespaces/index.ts
@@ -58,3 +58,4 @@ export * from "./applications-ns";
 export * from "./idp-ns";
 export * from "./api-resources-ns";
 export * from "./ai-ns";
+export * from "./impersonation-ns";